new gif and more bugs
