+++ src/pages/home/landing/landing.jsx
@@ -6,6 +6,7 @@ import "./landing.css"
 
 import imgHow from '../../../images/howToSearch.gif'
 import imgCreate from '../../../images/imgCreate.gif'
+import imgSave from "../../../images/saveFun.gif"
 
 // Section Code
 
@@ -32,6 +33,15 @@ function LandingPage(){
                 </div>
                 <div className="sectionGifs">
                     <img src={imgCreate} width={1000} />
+                </div>
+            </div>
+
+            <div className="sectionMain">
+                <div className="sectionTitle">
+                    <h1>Salve uma função</h1>
+                </div>
+                <div className="sectionGifs">
+                    <img src={imgSave} width={1000} />
                 </div>
             </div>
